backend/pyproject.toml
[project]
name = "citizen-support-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "livekit-agents[silero,turn-detector]>=1.0",
    "livekit-plugins-openai>=1.0",
    "livekit-plugins-groq>=1.0",
    "llama-index>=0.10",
    "llama-index-embeddings-openai>=0.1",
    "llama-index-readers-file>=0.1",
    "pypdf>=4.0",
    "python-docx>=1.0",
    "groq>=0.5",
    "openai>=1.0",
    "redis>=5.0",
    "fastapi>=0.100",
    "uvicorn>=0.23",
    "gunicorn>=21.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0",
    "pydantic-settings>=2.0",
    "aiofiles>=23.0",
    "pydantic-ai[openai]>=0.0.38",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
